feat : util 함수 파일 위치 이동
