+++ package.json
@@ -44,6 +44,7 @@
     "react-router-dom": "^6.20.0",
     "react-toastify": "^9.1.3",
     "rimraf": "^5.0.1",
+    "uuid": "^9.0.1",
     "webpack": "^5.88.2",
     "zustand": "^4.4.1"
   },
@@ -66,6 +67,7 @@
     "@types/react-dom": "^18.2.7",
     "@types/react-grid-layout": "^1.3.2",
     "@types/react-helmet": "^6.1.6",
+    "@types/uuid": "^9.0.7",
     "@types/webpack-env": "^1.18.1",
     "@typescript-eslint/eslint-plugin": "^6.6.0",
     "babel-loader": "^9.1.3",
