+++ src/components/SharePage/Blocks/PageBlock/PageBlock.module.scss
@@ -30,7 +30,7 @@
         margin-left: 1.25rem;
         h3 {
           @include flex(column, null);
-          font-size: $text-h-4;
+          font-size: $text-p;
           margin: 0;
         }
         p {
