+++ src/components/SharePage/DnDViewer/BlockDropDown.module.scss
@@ -1,3 +1,6 @@
+@import '/src/styles/variables.scss';
+@import '/src/styles/mixin.scss';
+
 .iconCover {
   display: flex;
   justify-content: center;
@@ -23,7 +26,6 @@
   text-align: start;
   border: none;
   background-color: transparent;
-  font-size: 14px !important;
   padding: 0px;
   cursor: pointer;
 }
@@ -34,7 +36,6 @@
   width: 100%;
   text-align: start;
   background-color: transparent;
-  font-size: 14px !important;
   padding: 0px;
   margin: 0px;
   cursor: pointer;
