+++ src/components/SharePage/DnDViewer/BlockDropDown.tsx
@@ -1,28 +1,23 @@
 import { BsThreeDotsVertical } from '@react-icons/all-files/bs/BsThreeDotsVertical'
 import { Switch } from 'antd'
 import { useMemo } from 'react'
-import { editTextBlockAPI } from '@/apis/blocks'
-import { editImageBlockAPI } from '@/apis/blocks/imageBlock'
-import { editLinkBlockAPI } from '@/apis/blocks/linkBlock'
-import { editGoogleMapBlockAPI } from '@/apis/blocks/mapBlock'
-import { editVideoBlockAPI } from '@/apis/blocks/videoBlock'
-import { editPageBlockAPI } from '@/apis/page'
 import { DropDownMenu } from '@/components/SharePage/DropDownMenu'
 import { DROPDOWN_TRIGGER_ICON_ID } from '@/constants'
-import { IMAGE, LINK, TEXT, BLOCK, MAP, PAGE, TEMPLATE, VIDEO, BLOCK_TO } from '@/constants/block'
+import { BLOCK, PAGE, TEMPLATE, BLOCK_TO } from '@/constants/block'
 import { type BlockBase, type BlockType, type BlockWith } from '@/models/block'
 import { useSharePageQuery } from '@/queries/useSharePageQuery'
 import { useBlockActionStore } from '@/store/blockAction'
-import { clip, toast } from '@/utils'
-import styles from './EditorDropDownMenu.module.scss'
+import { clip } from '@/utils'
+import { editBlockAPIByType, getUniqueDivier } from '@/utils/SharePage'
+import styles from './BlockDropDown.module.scss'
 
-interface EditorDropDownMenuProps {
+interface BlockDropDownProps {
   block: BlockBase<BlockType>
   onDelete: (pageId: string | undefined, block: BlockBase<BlockType>) => void
 }
-export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMenuProps) {
+export default function BlockDropDown({ block, onDelete }: BlockDropDownProps) {
   const { activeBlockId } = useBlockActionStore()
-  const { sharePage, pageId } = useSharePageQuery()
+  const { pageId } = useSharePageQuery()
   const { setOpenDrawer, setFormType, setDrawerMode, setBlockType } = useBlockActionStore()
 
   const items = useMemo(() => {
@@ -34,20 +29,13 @@ export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMe
             className={styles.switchBtn}
             defaultChecked={block.visible}
             onChange={() => {
-              // TODO : 이거 무슨 로직인지 확인하기
-              switchToggleAPIByBlockType(pageId, block)
-              for (let i = 0; i < sharePage.children.length; i++) {
-                if (sharePage.children[i].objectId === block.objectId) {
-                  sharePage.children[i].visible = !sharePage.children[i].visible
-                  break
-                }
-              }
+              editBlockAPIByType(pageId, block)
             }}
           />
         ),
         icon: '공개 설정'
       },
-      getUniqueDivierItem(`${block.objectId}-ViewerBlockBase-divier-1`),
+      getUniqueDivier(),
       {
         key: `${block.objectId}-view-block-2`,
         label: (
@@ -71,7 +59,7 @@ export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMe
     ]
 
     const TemplateSwitchItem = [
-      getUniqueDivierItem(`${block.objectId}-ViewerBlockBase-divier-2`),
+      getUniqueDivier(),
       {
         key: `${block.objectId}-view-block-3`,
         label: (
@@ -80,7 +68,7 @@ export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMe
           </button>
         )
       },
-      getUniqueDivierItem(`${block.objectId}-ViewerBlockBase-divier-3`),
+      getUniqueDivier(),
       {
         key: `${block.objectId}-view-block-4`,
         label: <button className={styles.kebobBtn}>템플릿 상세 설정</button>
@@ -88,14 +76,13 @@ export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMe
     ]
 
     const deleteItem = [
-      getUniqueDivierItem(`${block.objectId}-ViewerBlockBase-divier-4`),
+      getUniqueDivier(),
       {
         key: `${block.objectId}-view-block-5`,
         danger: true,
         label: <p className={styles.delBtn}>삭제하기</p>,
         onClick: () => {
           onDelete(pageId, block)
-          //   setConfirmModal(true)
         }
       }
     ]
@@ -120,34 +107,4 @@ export default function EditorDropDownMenu({ block, onDelete }: EditorDropDownMe
       </button>
     </DropDownMenu>
   )
-}
-
-function getUniqueDivierItem(key: string) {
-  return {
-    key,
-    type: 'divider'
-  }
-}
-
-function switchToggleAPIByBlockType(pageId: string | undefined, block: BlockBase<BlockType>) {
-  if (!pageId) return toast("잘못된 접근입니다. 'pageId'가 존재하지 않습니다.", 'failure')
-  const { objectId, visible } = block
-  switch (block.type) {
-    case TEXT:
-      return editTextBlockAPI(pageId, { objectId, visible: !visible })
-    case IMAGE:
-      return editImageBlockAPI(pageId, { objectId, visible: !visible })
-    case LINK:
-      return editLinkBlockAPI(pageId, { objectId, visible: !visible })
-    case PAGE:
-      return editPageBlockAPI(pageId, { visible: !visible })
-    case VIDEO:
-      return editVideoBlockAPI(pageId, { objectId, visible: !visible })
-    case MAP:
-      return editGoogleMapBlockAPI(pageId, { objectId, visible: !visible })
-    case TEMPLATE:
-      return toast('템플릿 공개 설정은 아직 지원하지 않습니다.', 'success')
-    default:
-      return toast('잘못된 입력입니다.', 'failure')
-  }
 }
