+++ src/components/SharePage/Forms/MapBlockForm/AddMapBlock.tsx
@@ -5,7 +5,7 @@ import { addMapBlockMutate } from '@/queries/mutates/mapBlockMutate'
 import { useSharePageQuery } from '@/queries/useSharePageQuery'
 import { useBlockActionStore } from '@/store/blockAction'
 import { useMapStore } from '@/store/map'
-import { getNextYOfLastBlock } from '@/utils'
+import { getNextYOfLastBlock } from '@/utils/SharePage'
 import { toast } from '@/utils/toast'
 import { useLoadGoogleMap } from '@/hooks/useGoogleMap'
 import { MapBlockForm } from './MapBlockForm'
