+++ src/components/SharePage/Forms/VideoBlockForm/AddVideoBlock.tsx
@@ -3,7 +3,7 @@ import { type AddVideoBlockBody } from '@/apis/blocks/videoBlock'
 import { addVideoBlockMutate } from '@/queries/mutates/videoBlockMutate'
 import { useSharePageQuery } from '@/queries/useSharePageQuery'
 import { useBlockActionStore } from '@/store/blockAction'
-import { getNextYOfLastBlock } from '@/utils'
+import { getNextYOfLastBlock } from '@/utils/SharePage'
 import Form from './Form'
 import styles from './VideoBlockForm.module.scss'
